Litte fix, create custom exception 'NotFoundException'

+++ src/main/java/com/example/traningsystem/dao/GroupRepository.java
@@ -3,6 +3,6 @@ package com.example.traningsystem.dao;
 import com.example.traningsystem.model.Groups;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupRepository extends JpaRepository<Groups, Integer> {
+public interface GroupRepository extends JpaRepository<Groups, Long> {
 
 }
