+++ src/main/java/com/example/traningsystem/service/ServiceTeacher.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ServiceTeacher {
     void save(Teacher teacher);
     List<Teacher> findAllTeachers();
-    Teacher findTeacherById(Integer id);
-    void deleteTeacherById(Integer id);
+    Teacher findTeacherById(Long id);
+    void deleteTeacherById(Long id);
     Teacher updateTeacher(Teacher teacher);
 }
